Add more explanatory material for readable stream.
This provides a gentler introduction to how the API works than immediately jumping into the internal algorithms. We first introduce the constructor, and then the consumption APIs, before showing the internal state machine (in diagram form). Only then do we get to the actual specification of each method and the relevant internal operations.

- Includes a state diagram (#29).
- Fixes #49 by adding some explanation.
- Some examples shortened and incorporated into the spec; others moved to the new Examples.md.