Unify names for readable and writable. Closes #47.
Due to the breakthrough idea in https://github.com/whatwg/streams/issues/44#issuecomment-32196072, we will be able to represent both duplex and transform streams as `{ in, out }` pairs of writable/readable streams. This means that we'll never need to have both readable and writable methods on the same object, and so we can reuse names. We no longer need suffixes, like `waitForReadable`/`waitForWritable`, or slightly-different names for concepts, like abort vs. dispose, or finish vs. close.

In sum:

- Readable streams' "abort" -> "dispose"
- Readable streams' "finish" -> "close"
- `readableState` and `writableState` -> `state`
- `waitForReadable()` and `waitForWritable()` -> `wait()`