Direct pipe hookup: fix for pull sources. (Reference implementation only)
This reestablishes the meaning of the [[pulling]] flag to mean "is the `pull` callback executing right now," so that it can be used to avoid infinitely-recursive pull calls.

The meaning of [[pulling]] has gotten a bit confused. It does two things right now: prevent "too many" calls to `pull` in a row, and avoid reentrant calls in the pipe use case. (Reentrant calls are not possible in the non-pipe use case, since those will necessarily be mediated through wait(). I think.) Notably you can remove the line that checks [[pulling]] inside [[callPull]] without any tests failing except the one that is explicitly checking that `pull` is not called multiple times.