Initial draft of benchmark suite (#120).
This is meant to allow us to assess the impact of a spec change on the performance of the algorithms via the reference implementation.

This initial draft contains a decent amount of dead code from abandoned ideas that will be cleaned up in the subsequent commit; however, I thought it'd be useful to have it in source control in case someone wants to use it later.