Port over the spec entirely to Bikeshed/HTML
It now lives on https://streams.spec.whatwg.org/, with subdirectories automatically generated for each branch and commit (but of course master goes in the root). Fixes #211.

The parts of the README that were somewhat incomplete, like the strategies or transfer stream or pipeTo algorithm, were ported over still-incomplete, so that still needs to be done. TeeStream was moved to a separate file as an experimental extension, like the binary one.

While doing this, added non-normative descriptions for all the methods and getters, and fleshed out a few more concepts in the Model section.

While doing this, found and fixed a few errors-not-being-caught bugs (#128) and made a few other normative tweaks. In total, they are:

- set `[[startedPromise]]` to `undefined` after it fulfills, to free up that memory.
- Catch errors computing the chunk size in `writableStream.write(chunk)`.
- Catch errors enqueuing a chunk in `writableStream.write(chunk)`. (These would occur if the returned size was NaN/+Infinity/-Infinity.)
- Stopped using pointless write records of `{[[type]]: "close", [[chunk]]: undefined, [[promise]]: undefined}` and instead just used `"close"`.