Fix one waiting → waiting ready signal per #258
As demonstrated in https://github.com/whatwg/streams/issues/258#issuecomment-68828539, there was a bug where acquiring and then releasing a reader would cause a stream's ready promise to fulfill, despite the stream staying in a waiting state both before and after the fulfillment. This commit fixes that, which is one of two waiting → waiting transitions discovered in that bug.