Add tests for the new source/sink error cases
The previous commits introduced many new places that we call in to user code, since we no longer validate that we are grabbing functions and then store them, but instead extract and invoke methods every time. This commit adds tests for a variety of pathological inputs, to ensure that the errors they throw are propagated or caught-and-exposed, as appropriate.