Fix a variety of algorithm mistakes
This fixes a variety of spec reference errors, where the wrong variable or abstract operation is used in the spec. It also aligns the reference implementation better with the spec, either by fixing spec errors that the reference implementation got correct, or changing variable names in the reference implementation to match the spec.

PR: https://github.com/whatwg/streams/pull/442