Mostly-editorial tweaks to align with newer ES algorithm styles
This explicitly annotates all abstract operation and method calls with ! or ? as appropriate, per the conventions introduced in semi-recent ES spec drafts for dealing with abrupt completions. That subsumes several of our previous idiosyncratic conventions, such as a special meaning for "Perform" or the "Call-with-rethrow" shorthand.

While here, a few other style changes were made, such as spacing in list and record literals, or capping all lines at 120 characters.

The following notable non-editorial changes were made:

- Renamed ReadableStreamHasReader to ReadableStreamHasDefaultReader
- Handled abrupt completions while allocating a new ArrayBuffer for the auto-allocation feature (due to out of memory errors)
- Moved the "this abstract operation can be called by other specifications" from ReadableStreamClose and ReadableStreamError to their counterparts for the ReadableStreamDefaultController, since now those two abstract operations are instead used as part of the internal controller interface
- Formalized tranfers and other byte manipulation via the new SameRealmTransfer abstract operation and the existing CopyDataBlockBytes and CloneArrayBuffer abstract operations
- Made some cleanups to ReadableByteStreamControllerPullInto
- Changed ReadableStreamBYOBRequestInvalidate to ReadableByteStreamControllerInvalidateBYOBRequest, which contains all the logic which previously was redundantly included at every call site