Update reference implementation for Node v6
This is largely a matter of adding 'use strict' and converting imports into requires. With those changes made, we can rip out the Traceur build infrastructure.

This also removes the browser test build step, as the web platform tests project takes care of that use case now. Similarly, this removes the benchmark runner, as the benchmark was not being run regularly and was unmaintained in light of recent spec changes; if we want to pick that project back up we'll take a different approach.

Part of #445, although the broken tests are not yet fixed.