Update the push source byte stream example
This fixes #489, wherein it was noted that the EventTarget-based example did not work unless the stream drained the underlying source of data each time. We took this as an opportunity to fix up the example in general, as it made a number of problematic assumptions (e.g. that TCP sockets would lose data, or that there was a way to directly manipulate the TCP window size). The new example omits backpressure signaling, in favor of simply wrapping a UDP socket, since this allows us to illustrate the use of the stream's internal queue to prevent data loss.