Convert some general tests for ws into wpt
This commit is for the issue #492.

wpt-runner invocation is tweaked by domenic.

We'll use the to-upstream-wpts/ directory for placing converted wpt files until it's ready for upstreaming them to the w3c repo.

All the tests under the directory are run automatically when we run "npm run wpt" in addition to the ones already moved to the w3c repo.