Port constructor tests to web-platform-tests (#524)
Move tests relating to the WritableStream constructor and
newly-constructed objects to the testharness.js framework.

Only syntactic and style changes have been made in this CL. There are no
semantic changes. except that the test now checks that
WritableStreamDefaultWriter properties are defined, rather than truthy. Also, 
where the original test verified that a promise resolved, the ported version just
returns the promise and lets the test harness check it.