Port close() tests to web-platform-tests
Move tests relating to the WritableStream sink's close() method to the
testharness.js framework.

To reduce the risk of functional changes, changes to the tests are
intentionally minimal. The only kinds of changes are:

* Use testharness.js syntax
* Match web-platform-tests style

async_test was used where necessary to retain the existing
semantics. It will be changed to promise_test in a separate commit.