Port test that sink methods are called as methods
Port the test 'WritableStream should call underlying sink methods as
methods' to web-platform-tests. The tape test relied on assertions
working after the test function had completed. The new test creates a
promise for each method, then fulfills it with the value of 'this'. The
main logic then verifies that each promise is fulfilled with the correct
value of 'this'.