Add tests for WritableStreamDefaultWriter and Controller constructors
The WritableStreamDefaultWriter and WritableStreamDefaultController
constructors are not exported. However they can be accessed fron the
"constructor" property of the relevant objects.

The WritableStreamDefaultController constructor should always throw an
exception if used explicitly. Verify that it does.

WritableStreamDefaultWriter can be constructed with an unlocked stream
and should then work just like a writer returned from
getWriter(). Verify that it does.