Add done() to the end of w-p-t JavaScript files
When running a web-platform-test in a Worker, the JavaScript file must end with a done() statement so that testharness.js knows that parsing is complete. The extra done() statement is a no-op when running directly in the page or via the wpt-runner.

Add done() statements to all relevant JavaScript files.