rework recording streams helper
This is the first building block towards being able to record
the events of multiple streams in a single array for future
backpressure tests.

Maintains the old, simpler API which is easy to use with
assert_array_equals, which doesn't do a deep-compare of objects.

Also adds a .startPromise helper, since the recording code
now hooks into the promises in order to record when they're
fulfilled or rejected.