Invoking methods should not call .apply()
The specification documents InvokeOrNoop() and friends as using the
Ecmascript Call() abstract operation. The reference implementation was
calling .apply() on the method instead. Fix it to use
Function.prototype.apply.call() instead.

Also add tests to verify that .apply() and .call() on the underlyingSink
and transformer methods are not called.