Test that implementation does not call .apply() or .call() (#613)
The specification documents InvokeOrNoop() and friends as using the
Ecmascript Call() abstract operation. It is important that this is not
implemented via apply() or call() methods on the function object as this
would be visible to Javascript if they were overridden. Add tests to
verify this for TransformStream and WritableStream.

Also modify the reference implementation to pass the new tests.