Make an abort() not abort a sink write()
Previously, if abort() was called while a sink write() was in progress,
the promise returned from WritableStream.write() would be rejected.

This behaviour was surprising, since the rejection would happen even if
the underlying write succeeded.

Change the promise returned by WritableStream.write() to always reflect
the success or failure of the underlyingSink.write() that has started.