Reject the correct writer.closed promise on errored release
Prior to #619, abort() would reject the closed promise immediately. Now it waits for queued sink operations to finish. This means that there can be a window when the stream is errored but the closed promise has not been rejected. If releaseLock() was called during the window it would incorrectly create a new closed promise on the assumption it was already rejected.

Instead, when an abort() is pending, reject the promise rather than creating a new one.