WritableStream: Fix the interface boundary in the closure processing
Factor out rejection handling logic from WritableStreamDefaultControllerProcessClose()

The logic should be factored out as an operation exposed by the WritableStream to controllers.

This also improves readability of WritableStreamDefaultControllerProcessClose() slightly by making it easier to see that all the promises are correctly processed in both of the fulfillment callback and rejection callback.

The assertion at the top of WritableStreamFinishClose() has been removed because the assertion in the else clause of the following if-block is enough.

Closes #639