Meta: add a service worker to the spec
This allows the spec to be viewed offline.

We use a cache-then-network strategy for all subresources, but a network-then-cache strategy for the main document, since it is imperative you see the latest version of living standards. In the future we may explore more complicated ways of speeding up the main resource fetch, e.g. https://github.com/whatwg/streams/pull/637#issuecomment-270391258.