Do not call sink abort() while sink start() is running
Previously, calling abort() on a WritableStream or Writer would cause the sink
abort() to be called immediately if the sink start() was still executing. This
was inconsistent with the way write() and close() work and could be a cause of
bugs.

Postpone calling sink abort() until start() is complete. If start() errors, do
not call abort() at all.

Fixes #683.